Minor improvements to setup. Included example data conversion notebook.

+++ .devcontainer/setupData.sh
@@ -1,3 +1,9 @@
-kaggle datasets download -d dannielr/marvel-superheroes
-unzip marvel-superheroes.zip -d ./data/marvel
-rm marvel-superheroes.zip 
+
+FOLDER=data
+if test -e "$FOLDER"; then
+  echo "Data already downloaded"
+else
+  kaggle datasets download -d dannielr/marvel-superheroes
+  unzip marvel-superheroes.zip -d ./data/marvel
+  rm marvel-superheroes.zip 
+fi
